Сделал миддлвару для для валидации данных при создании и обновлении контакта.
