+++ middlewares/index.js
@@ -1,0 +1,5 @@
+const validation = require('./validation');
+
+module.exports = {
+  validation,
+};
