+++ schemas/contacts/addContact.js
@@ -1,0 +1,11 @@
+const Joi = require('joi');
+
+const schemaCreateContact = Joi.object({
+  name: Joi.string().alphanum().required(),
+  email: Joi.string().email().required(),
+  phone: Joi.string()
+    .regex(/^(\+)?(\(\d{2,3}\) ?\d|\d)(([ \-]?\d)|( ?\(\d{2,3}\) ?)){5,12}\d$/)
+    .required(),
+});
+
+module.exports = schemaCreateContact;
