+++ schemas/contacts/index.js
@@ -1,0 +1,7 @@
+const schemaCreateContact = require('./addContact');
+const schemaUpdateContact = require('./updateContact');
+
+module.exports = {
+  schemaCreateContact,
+  schemaUpdateContact,
+};
