+++ schemas/index.js
@@ -1,0 +1,5 @@
+const contactsSchema = require('./contacts');
+
+module.exports = {
+  contactsSchema,
+};
